backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.5
anthropic>=0.39
slack-sdk>=3.27
python-dotenv>=1.0
